MPI+OpenMP+CUDA histogram
Calculation of a histogram of a large array containing values from 0 to 255. As a constraint, the architecture of cuda contains 10 blocks of 20 threads. Designed for 2 processes/devices on same LAN. Assumption of OpenMP is 4 threads to utilize the resources of local machine for maximum capacity, since my local machine contains 4 cores.